+++ CHANGELOG.md
@@ -1,3 +1,52 @@
+## [1.184.7](https://github.com/SocialGouv/mano/compare/v1.184.6...v1.184.7) (2022-12-06)
+
+
+### Bug Fixes
+
+* enable accueil de jour ([#1047](https://github.com/SocialGouv/mano/issues/1047)) ([a8414ad](https://github.com/SocialGouv/mano/commit/a8414ad39e3e62d6826292c5fdb7f94b23e4c9b0))
+
+## [1.184.6](https://github.com/SocialGouv/mano/compare/v1.184.5...v1.184.6) (2022-12-06)
+
+
+### Bug Fixes
+
+* name in consultations ([#1046](https://github.com/SocialGouv/mano/issues/1046)) ([dc2ce80](https://github.com/SocialGouv/mano/commit/dc2ce8011e7543b80cbf10e490b89d2a96c61ebc))
+
+## [1.184.5](https://github.com/SocialGouv/mano/compare/v1.184.4...v1.184.5) (2022-12-06)
+
+
+### Bug Fixes
+
+* **dashboard:** weekly dashboard display categories twice ([03d430e](https://github.com/SocialGouv/mano/commit/03d430edc5769b730c9b447caa95380e13605dc5))
+
+## [1.184.4](https://github.com/SocialGouv/mano/compare/v1.184.3...v1.184.4) (2022-12-06)
+
+
+### Bug Fixes
+
+* stop show lastLoginAt when never login ([f836d9c](https://github.com/SocialGouv/mano/commit/f836d9cd18cbbcc411c708c7e5d2d2fda27d82c8))
+
+## [1.184.3](https://github.com/SocialGouv/mano/compare/v1.184.2...v1.184.3) (2022-12-06)
+
+
+### Bug Fixes
+
+* rencontres fullscreen + default date for passage ([#1044](https://github.com/SocialGouv/mano/issues/1044)) ([07d315c](https://github.com/SocialGouv/mano/commit/07d315c369849c2744585e1d79707183e0bad82d))
+
+## [1.184.2](https://github.com/SocialGouv/mano/compare/v1.184.1...v1.184.2) (2022-12-05)
+
+
+### Bug Fixes
+
+* zod on actions and consultations ([#1041](https://github.com/SocialGouv/mano/issues/1041)) ([11441ec](https://github.com/SocialGouv/mano/commit/11441ec7917bbd272b01e87e61a9f424988cb0b4))
+
+## [1.184.1](https://github.com/SocialGouv/mano/compare/v1.184.0...v1.184.1) (2022-12-05)
+
+
+### Bug Fixes
+
+* filter type number MANO-5D ([cb49620](https://github.com/SocialGouv/mano/commit/cb49620c15477f0001c0c881d9662bcfabebae7c))
+
 # [1.184.0](https://github.com/SocialGouv/mano/compare/v1.183.0...v1.184.0) (2022-12-05)